Fix type warning in map handler and editor

+++ mapEditor/src/panel/ai/AIPanelWidget.cpp
@@ -82,7 +82,7 @@ namespace urchin {
     void AIPanelWidget::aiChanged() {
         if (!disableAIEvent) {
             std::shared_ptr<NavMeshAgent> navMeshAgent = std::make_shared<NavMeshAgent>(static_cast<float>(agentHeight->value()), static_cast<float>(agentRadius->value()));
-            navMeshAgent->setMaxSlope(AngleConverter<float>::toRadian(maxSlope->value()));
+            navMeshAgent->setMaxSlope(AngleConverter<float>::toRadian((float)maxSlope->value()));
             navMeshAgent->setJumpDistance(static_cast<float>(jumpDistance->value()));
 
             aiController->updateSceneAI(navMeshAgent);
